% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fart.R
\name{fart}
\alias{fart}
\title{Embarrass the user with a fart noise}
\usage{
fart(takes = NA)
}
\arguments{
\item{takes}{: Takes no parameter, NA by default}
}
\description{
A short natural noise is played which is useful if you want to get notified,
for example, when the script encounters an error.
It is recommended to wrap this function on top of your script like so:
options(error = function() {fart(takes= NA)})
This way whole namespace becomes visible; otherwise,
you can play sound wherever you wish, like so: fart(takes= NA)
}
\examples{
\dontrun{fart(takes= NA)}
\dontrun{options(error = function() {fart(takes= NA)})}

}
